Refactor login screens into multiple components, create confirmation code control
